arreglos pvs

+++ TP2-SO-2025/Userland/Shell/commands/test_prio.c
@@ -21,7 +21,8 @@ void zero_to_max() {
   uint64_t value = 0;
   while (value++ != max_value);
 
-  printf("PROCESS %d DONE!\n", my_getpid());
+  /* my_getpid() returns a pid type; print as int for this printf implementation */
+  printf("PROCESS %d DONE!\n", (int)my_getpid());
 }
 
 uint64_t _test_prio(uint64_t argc, char *argv[]) {
@@ -30,7 +31,8 @@ uint64_t _test_prio(uint64_t argc, char *argv[]) {
   char *ztm_argv[] = {0};
   uint64_t i;
 
-  printf("argc: %d\n", argc);
+  /* argc is a 64-bit value here; project printf supports %d, so cast to int */
+  printf("argc: %d\n", (int)argc);
   if (argc != 2)
     return -1;
 
@@ -54,7 +56,7 @@ uint64_t _test_prio(uint64_t argc, char *argv[]) {
   for (i = 0; i < TOTAL_PROCESSES; i++) {
     pids[i] = my_create_process(zero_to_max, 0, ztm_argv);
     my_nice(pids[i], prio[i]);
-    printf("PROCESS %d NEW PRIORITY: %d\n", pids[i], prio[i]);
+    printf("PROCESS %d NEW PRIORITY: %d\n", (int)pids[i], (int)prio[i]);
   }
 
   // Expect the priorities to take effect
@@ -68,7 +70,7 @@ uint64_t _test_prio(uint64_t argc, char *argv[]) {
     pids[i] = my_create_process(zero_to_max, 0, ztm_argv);
     my_block(pids[i]);
     my_nice(pids[i], prio[i]);
-    printf("PROCESS %d NEW PRIORITY: %d\n", pids[i], prio[i]);
+    printf("PROCESS %d NEW PRIORITY: %d\n", (int)pids[i], (int)prio[i]);
   }
 
   for (i = 0; i < TOTAL_PROCESSES; i++)
